lecture-06/restintro/restintro/serializer.py
from rest_framework import serializers
from restintro.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'name', 'email', 'age']

